refactor: Paper Header 리팩토링
