+++ jsconfig.paths.json
@@ -3,6 +3,7 @@
     "baseUrl": "src",
     "paths": {
       "@Components/*": ["./components/*"],
+      "@Pages/*": ["./pages/*"],
       "@List/*": ["./pages/list/*"],
       "@Paper/*": ["./pages/paper/*"],
       "@Form/*": ["./pages/form/*"],
