--- src/Login.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function Login() {
-  return <div>Test Login</div>;
-}
-
-export default Login;
