+++ src/components/PaperHeader/EmojiChoice.jsx
@@ -1,9 +1,12 @@
 import EmojiPicker from 'emoji-picker-react';
-import React from 'react';
+import React, { useContext } from 'react';
 import { useParams } from 'react-router-dom';
-import createReactions from '../api/api';
+import { PaperHeaderContext } from '@Components/PaperHeader/context/PaperHeaderContext';
+import createReactions from '@Pages/paper/api';
 
-function EmojiChoice({ onUpdateEmoji }) {
+function EmojiChoice() {
+  const { updateEmoji, emojis } = useContext(PaperHeaderContext);
+
   const { recipientsId } = useParams();
   const handleEmojiClick = async (emoji) => {
     try {
@@ -12,8 +15,9 @@ function EmojiChoice({ onUpdateEmoji }) {
         emoji: emoji.emoji,
       };
 
-      const response = await createReactions(recipientsId, data);
-      onUpdateEmoji(emoji.emoji);
+      await createReactions(recipientsId, data);
+      updateEmoji(emoji.emoji);
+      console.log(emojis);
     } catch (error) {
       console.error('이모지 전송 실패:', error);
     }
