+++ src/components/PaperHeader/EmojiChoice.style.js
