+++ src/components/PaperHeader/EmojiCountBtn.style.js
